requirements.txt
einops>=0.3.0
numpy>=1.19.5
torchvision>=0.10.0
antialiased_cnns
axial_attention
pytorch_msssim
huggingface_hub